Move the budget items logic to its own file
